Переносит логику в контроллер
